[mod]StorageRecord格式調整

+++ app/src/main/java/com/lhr/water/repository/FormRepository.kt
@@ -137,7 +137,7 @@ class FormRepository(context: Context) {
 
         // 篩選後的List中數量加總
         for (storageContentEntity in filteredList) {
-            totalQuantity += storageContentEntity.quantity
+            totalQuantity += storageContentEntity.quantity.toInt()
         }
         return totalQuantity
     }
